refactor: 배치 패키지 하위 클래스 구조 변경
