+++ src/main/kotlin/com/asap/asapbackend/batch/timetable/dto/ElsTimetable.kt
@@ -1,4 +1,4 @@
-package com.asap.asapbackend.batch.timetable.model
+package com.asap.asapbackend.batch.timetable.dto
 
 data class ElsTimetable(
         var elsTimetable: List<Row>
