+++ src/main/kotlin/com/asap/asapbackend/batch/timetable/dto/Head.kt
@@ -1,4 +1,4 @@
-package com.asap.asapbackend.batch.timetable.model
+package com.asap.asapbackend.batch.timetable.dto
 
 data class Head(
         val list_total_count: Int?,
