+++ src/main/kotlin/com/asap/asapbackend/batch/timetable/dto/Result.kt
@@ -1,4 +1,4 @@
-package com.asap.asapbackend.batch.timetable.model
+package com.asap.asapbackend.batch.timetable.dto
 
 data class Result(
         val CODE: String?,
