+++ src/main/kotlin/com/asap/asapbackend/batch/timetable/dto/Timetable.kt
@@ -1,0 +1,6 @@
+package com.asap.asapbackend.batch.timetable.dto
+
+data class Timetable(
+    val head: List<Head>?,
+    val row: List<Row>?
+)
