+++ src/main/kotlin/com/asap/asapbackend/batch/timetable/dto/TimetableResponse.kt
@@ -1,4 +1,4 @@
-package com.asap.asapbackend.batch.timetable.model
+package com.asap.asapbackend.batch.timetable.dto
 
 data class TimetableResponse (
         val day: String?,
