+++ src/main/kotlin/com/asap/asapbackend/batch/timetable/exception/NotFoundTimetableException.kt
