--- src/main/kotlin/com/asap/asapbackend/batch/timetable/model/Timetable.kt
@@ -1,6 +1,0 @@
-package com.asap.asapbackend.batch.timetable.model
-
-data class Timetable(
-        val head: List<Head>?,
-        val row: List<Row>?
-)
